inline: turn on with_enter_leave
any nextstate in the cv body will reset the stack to a wrong value
(oldsp from the upper block), so we cannot just leave out enter/leave,
we need this oldsp.

but leave will just reset the sp to the old value, and thereby undo
all the cv sideeffects, pushing the return values onto our stack.