Switch some core modules to XSLoader
RT 132080: Remove perl 5.006 compatibilities with DynaLoader and use
XSLoader directly.

The traditional boiler plate to use XSLoader for Perl > 5.006 or
DynaLoader, does not make sense for core modules in perl 5.28+.

cperl: i.e. explicitly break cpan for <5.6. which is okay with me, but
should be stated as such. This is not only a core module for 5.28, it is
a critical cpan module for all perls and thus needs version guards in
it's Makefile.PL